backend/requirements.txt
python-dotenv==1.0.1
Flask==3.0.2
Flask-HTTPAuth==4.8.0
waitress==3.0.0
argon2-cffi==23.1.0
flask-jwt-extended==4.6.0
